Sometimes (often, even) when writing tools one encounters the situation in which a term must be handled of unknown origin and encoding.  And to use trm2bar/baf2trm, the storage format had to be known, on penalty of coredumps.
I've made `baffle' a little more less picky about its input: it'll happily
read binary as well as text format input in both modes of operations,
writing the requested format.

This way, it can be used to handle terms of unknown encoding.  As a
reflection of this change, the tiny tool `trmcat' is added, that
writes arbitrary format terms as text output.